Created models for representing environment variables.
A value instance consists of string content and a boolean state. The state
is used to indicate whether the value is applied.

A variable instance consists of a string name, a boolean state, a type,
and a collection of values. The type is used to indicate whether the
variable is a system or user environment variable. Once again, the state
is used to indicate whether the value is applied.