Created MainWindowViewModel class.
Defines the view model for the main window and provides properties to (1)
get the collection of system and user variables separately, (2) retieve
the values associated with a particular variable, (3) import and export
a system environment, and (4) apply or revert a system environment.

Uses DelegateCommands to define the action and whether or not it can be
performed.