Created MainWindow view.
Displays a menu with options to import or export an environment
configuration.

Has two tabs: one for system variables, and the other for user variables.
Each pane has two data grids: one for displaying the variable name and
state, and the other for displaying the value state and content.

Unfortunately hidden columns with a default value specified do not
participate in the creation of a new item, so the OnInitializingNewItem
handlers were a workaround to set the type of variable correctly.